Changed seed from long to int since JSONAdapter / FIleAdapter only supports int

+++ raylibBackend/inc/Noise.h
@@ -9,7 +9,7 @@ namespace Noise {
 	struct noise_layer_settings;
 
 	struct noise_settings {
-		long seed;
+		int seed;
 		std::vector<noise_layer_settings> noiseLayerSettings;
 	};
